Refactor top-level modules
