+++ aeon/analyze/plotting.py
