+++ docs/examples/exp01_patchmovies.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 from aeon.analyze.movies import *
 from aeon.analyze.patches import *
-from aeon.util.plotting import *
+from aeon.analyze.plotting import *
 import aeon.io.api as aeon
 import aeon.io.video as video
 import matplotlib.pyplot as plt
